requirements.txt
# Usage: pip install -r requirements.txt

#./Required/Dlib/dlib-19.22.99-cp310-cp310-win_amd64.whl
cmake
pyttsx3>=2.90
edge_tts>=6.1.8
asyncio>=3.4.3
SpeechRecognition>=3.9.0
playsound==1.2.2
PyAudio>=0.2.12
PyAutoGUI>=0.9.53
#selenium>=4.4.3
pyserial>=3.5
speedtest-cli>=2.1.3
wikipedia>=1.4.0
pywhatkit>=5.4
requests>=2.28.1
#python-vlc>=3.0.16120
opencv-python>=4.6.0.66
numpy>=1.18.5
mediapipe
pystray
cvzone>=1.5.6
face-recognition>=1.3.0

#python -m PyInstaller jarvis-english.py

